iot-devices/requirements.txt
paho-mqtt
python-dotenv
pika
requests
psutil
